+++ TeePod/modules/Home/HomeEntity.swift
@@ -8,5 +8,3 @@
 
 import Foundation
 import SwiftUI
-
-struct HomeEntiry {}
